requirements.txt
PyOpenAL
PyOgg
Pygame-menu
pygame